+++ validate/src/request.rs
@@ -25,10 +25,10 @@ pub const GET_CHANNEL_MESSAGES_LIMIT_MAX: u8 = 100;
 pub const GET_CHANNEL_MESSAGES_LIMIT_MIN: u8 = 1;
 
 /// Maximum amount of guilds to get.
-pub const GET_CURRENT_USER_GUILDS_LIMIT_MAX: u64 = 200;
+pub const GET_CURRENT_USER_GUILDS_LIMIT_MAX: u8 = 200;
 
 /// Minimum amount of guilds to get.
-pub const GET_CURRENT_USER_GUILDS_LIMIT_MIN: u64 = 1;
+pub const GET_CURRENT_USER_GUILDS_LIMIT_MIN: u8 = 1;
 
 /// Maximum amount of audit log entries to list.
 pub const GET_GUILD_AUDIT_LOG_LIMIT_MAX: u8 = 100;
@@ -356,7 +356,7 @@ pub enum ValidationErrorType {
     /// Provided get current user guilds limit was invalid.
     GetCurrentUserGuilds {
         /// Invalid limit.
-        limit: u64,
+        limit: u8,
     },
     /// Provided get guild audit log limit was invalid.
     GetGuildAuditLog {
@@ -546,7 +546,7 @@ pub const fn get_channel_messages_limit(limit: u8) -> Result<(), ValidationError
 ///
 /// [`GetCurrentUserGuilds`]: ValidationErrorType::GetCurrentUserGuilds
 /// [this documentation entry]: https://discord.com/developers/docs/resources/user#get-current-user-guilds
-pub const fn get_current_user_guilds_limit(limit: u64) -> Result<(), ValidationError> {
+pub const fn get_current_user_guilds_limit(limit: u8) -> Result<(), ValidationError> {
     if limit >= GET_CURRENT_USER_GUILDS_LIMIT_MIN && limit <= GET_CURRENT_USER_GUILDS_LIMIT_MAX {
         Ok(())
     } else {
